Bug fixes and admin profile fixes

+++ controllers/admin/productController.js
@@ -17,29 +17,33 @@ var productController = {
             if(req.session.userId) loggedIn = true;
             else loggedIn = false;
 
-
-            if(result != null) {
-                const details = {
-                    result,
-                    title: "Admin | Admin Products",
-                    loggedIn: loggedIn,
-                    userId: req.session.userId,
-                    username: req.session.adminUsername,
-                    error: null,
-                    path
-                };
-                res.render('admin/admin-product', details);
+            if(loggedIn){
+                if(result != null) {
+                    const details = {
+                        result,
+                        title: "Admin | Admin Products",
+                        loggedIn: loggedIn,
+                        userId: req.session.userId,
+                        username: req.session.adminUsername,
+                        error: null,
+                        path
+                    };
+                    res.render('admin/admin-product', details);
+                }
+                else {
+                    const details = {
+                        result,
+                        title: "Baked Goods | Error 404",
+                        loggedIn: loggedIn,
+                        userId: req.session.userId,
+                        username: req.session.adminUsername,
+                        error: "404: Page not Found."
+                    };
+                    res.render('admin/admin-error', details);
+                }
             }
             else {
-                const details = {
-                    result,
-                    title: "Baked Goods | Error 404",
-                    loggedIn: loggedIn,
-                    userId: req.session.userId,
-                    username: req.session.adminUsername,
-                    error: "404: Page not Found."
-                };
-                res.render('admin/admin-error', details);
+                res.redirect('/admin/adminLogin');
             }
         });
     },
@@ -53,28 +57,33 @@ var productController = {
             if(req.session.userId) loggedIn = true;
             else loggedIn = false;
 
-            if(result != null) {
-                req.session.pName = result.name;
-                const details = {
-                    result,
-                    title: "Baked Goods | " + result.name,
-                    loggedIn: loggedIn,
-                    userId: req.session.userId,
-                    username: req.session.adminUsername,
-                    error: null
-                };
+            if(loggedIn){
+                if(result != null) {
+                    req.session.pName = result.name;
+                    const details = {
+                        result,
+                        title: "Baked Goods | " + result.name,
+                        loggedIn: loggedIn,
+                        userId: req.session.userId,
+                        username: req.session.adminUsername,
+                        error: null
+                    };
 
-                res.render('admin/admin-edit-product', details);
-            } else {
-                const details = {
-                    result,
-                    title: "Baked Goods | Error 404",
-                    loggedIn: loggedIn,
-                    userId: req.session.userId,
-                    username: req.session.adminUsername,
-                    error: "404: Page not Found."
-                };
-                res.render('admin/admin-error', details);
+                    res.render('admin/admin-edit-product', details);
+                } else {
+                    const details = {
+                        result,
+                        title: "Baked Goods | Error 404",
+                        loggedIn: loggedIn,
+                        userId: req.session.userId,
+                        username: req.session.adminUsername,
+                        error: "404: Page not Found."
+                    };
+                    res.render('admin/admin-error', details);
+                }
+            }
+            else{
+                res.redirect('/admin/adminLogin');
             }
         });
     },
@@ -85,15 +94,21 @@ var productController = {
         if(req.session.userId) loggedIn = true;
         else loggedIn = false;
 
-        const details = {
-            title: "Admin | User Accounts",
-            loggedIn: loggedIn,
-            userId: req.session.userId,
-            username: req.session.adminUsername,
-            error: null,
-            path
-        };
-        res.render('admin/admin-add-product', details)
+        if(loggedIn){
+            const details = {
+                title: "Admin | User Accounts",
+                loggedIn: loggedIn,
+                userId: req.session.userId,
+                username: req.session.adminUsername,
+                error: null,
+                path
+            };
+            res.render('admin/admin-add-product', details)
+        }
+        else{
+            res.redirect('/admin/adminLogin');
+        }
+
     },
 
     postProduct : function(req,res){
@@ -113,40 +128,47 @@ var productController = {
                 if(req.body.size3 != null)
                     sizes.push({size: req.body.size3, price: req.body.price3});
             }
+            var loggedIn = false;
+            if(req.session.userId) loggedIn = true;
 
-            database.addOne(product, {
-                productId: prodID,
-                name: Name,
-                description: Description,
-                sizes: sizes,
-                image: Image
-            }, function(result){
-                if(result != null){
-                    const details = {
-                        result,
-                        title: "Baked Goods | " + Name,
-                        headertitle: "Successfully added " + Name,
-                        loggedIn: true,
-                        userId: req.session.userId,
-                        username: req.session.adminUsername,
-                        line1: Name + " has been successfully added to the list of products!",
-                        line2: "You can view the list of products through the Product Management Tab",
-                        link: "/admin/admin-product"
-                    };
-                    res.render('admin/admin-success', details);
-                }
-                else{
-                    const details = {
-                        result,
-                        title: "Baked Goods | Error",
-                        loggedIn: true,
-                        userId: req.session.userId,
-                        username: req.session.adminUsername,
-                        error: "Oops! something went wrong with adding " + Name,
-                    };
-                    res.render('admin/admin-error', details);
-                }
-            });
+            if(loggedIn){
+                database.addOne(product, {
+                    productId: prodID,
+                    name: Name,
+                    description: Description,
+                    sizes: sizes,
+                    image: Image
+                }, function(result){
+                    if(result != null){
+                        const details = {
+                            result,
+                            title: "Baked Goods | " + Name,
+                            headertitle: "Successfully added " + Name,
+                            loggedIn: true,
+                            userId: req.session.userId,
+                            username: req.session.adminUsername,
+                            line1: Name + " has been successfully added to the list of products!",
+                            line2: "You can view the list of products through the Product Management Tab",
+                            link: "/admin/admin-product"
+                        };
+                        res.render('admin/admin-success', details);
+                    }
+                    else{
+                        const details = {
+                            result,
+                            title: "Baked Goods | Error",
+                            loggedIn: true,
+                            userId: req.session.userId,
+                            username: req.session.adminUsername,
+                            error: "Oops! something went wrong with adding " + Name,
+                        };
+                        res.render('admin/admin-error', details);
+                    }
+                });
+            }
+            else{
+                res.redirect('/admin/adminLogin');
+            }
     },
 
     postEdit : function(req, res){
@@ -188,38 +210,42 @@ var productController = {
                 }
         }
 
+        if(loggedIn){
+            database.updateOne(product, filter, update, function(flag) {
+                const projection = '';
+                const query = {_id: req.params.id}
+                database.findOne(product, query, projection, function(result) {
+                    if(result != null) {
 
-        database.updateOne(product, filter, update, function(flag) {
-            const projection = '';
-            const query = {_id: req.params.id}
-            database.findOne(product, query, projection, function(result) {
-                if(result != null) {
-
-                    const details = {
-                        result,
-                        title: "Baked Goods | " + Name,
-                        headertitle: "Successfully Updated " + Name,
-                        loggedIn: loggedIn,
-                        userId: req.session.userId,
-                        username: req.session.adminUsername,
-                        line1: "Data for " + Name + " been successfully updated!",
-                        line2: "You can view the list of products through the Product Management Tab",
-                        link: "/admin/admin-product"
-                    };
-                    res.render('admin/admin-success', details);
-                } else {
-                    const details = {
-                        result,
-                        title: "Baked Goods | Error",
-                        loggedIn: loggedIn,
-                        userId: req.session.userId,
-                        username: req.session.adminUsername,
-                        error: "Oops! something went wrong with updating" + Name
-                    };
-                    res.render('admin/admin-error', details);
-                }
+                        const details = {
+                            result,
+                            title: "Baked Goods | " + Name,
+                            headertitle: "Successfully Updated " + Name,
+                            loggedIn: loggedIn,
+                            userId: req.session.userId,
+                            username: req.session.adminUsername,
+                            line1: "Data for " + Name + " been successfully updated!",
+                            line2: "You can view the list of products through the Product Management Tab",
+                            link: "/admin/admin-product"
+                        };
+                        res.render('admin/admin-success', details);
+                    } else {
+                        const details = {
+                            result,
+                            title: "Baked Goods | Error",
+                            loggedIn: loggedIn,
+                            userId: req.session.userId,
+                            username: req.session.adminUsername,
+                            error: "Oops! something went wrong with updating" + Name
+                        };
+                        res.render('admin/admin-error', details);
+                    }
+                });
             });
-        });
+        }
+        else {
+            res.redirect('/admin/adminLogin');
+        }
     },
 
     deleteProduct: function(req, res) {
@@ -228,43 +254,47 @@ var productController = {
         var loggedIn = false;
         var projection = 'image username';
         if(req.session.userId) loggedIn = true;
-        database.findOne(product, filter, projection,function(result){
-            var img = result.image;
-            var Name = result.name;
 
-            if(img != null && fs.existsSync('public/img/products/'+ img))
-                fs.unlinkSync('public/img/products/'+ img);
+        if(loggedIn){
+            database.findOne(product, filter, projection,function(result){
+                var img = result.image;
+                var Name = result.name;
 
-            database.deleteOne(product,filter,  function(result){
+                if(img != null && fs.existsSync('public/img/products/'+ img))
+                    fs.unlinkSync('public/img/products/'+ img);
 
-                if(result!= null) {
-                    const details = {
-                        title: "Baked Goods | Delete Successful",
-                        headertitle: "Successfully Deleted Product",
-                        loggedIn: loggedIn,
-                        userId: req.session.userId,
-                        username: req.session.adminUsername,
-                        line1: Name + " has been removed from the list of products.",
-                        line2: "You can view the list of products through the Product Management Tab",
-                        link: "/admin/admin-product"
-                    };
-                    res.render('admin/admin-success', details);
-                }
-                else{
-                    const details = {
-                        result,
-                        title: "Baked Goods | Error",
-                        loggedIn: loggedIn,
-                        userId: req.session.userId,
-                        username: req.session.adminUsername,
-                        error: "Oops! something went wrong in deleting" + Name
-                    };
-                    res.render('admin/admin-error', details);
-                }
-            });
-        })
+                database.deleteOne(product,filter,  function(result){
 
-
+                    if(result!= null) {
+                        const details = {
+                            title: "Baked Goods | Delete Successful",
+                            headertitle: "Successfully Deleted Product",
+                            loggedIn: loggedIn,
+                            userId: req.session.userId,
+                            username: req.session.adminUsername,
+                            line1: Name + " has been removed from the list of products.",
+                            line2: "You can view the list of products through the Product Management Tab",
+                            link: "/admin/admin-product"
+                        };
+                        res.render('admin/admin-success', details);
+                    }
+                    else{
+                        const details = {
+                            result,
+                            title: "Baked Goods | Error",
+                            loggedIn: loggedIn,
+                            userId: req.session.userId,
+                            username: req.session.adminUsername,
+                            error: "Oops! something went wrong in deleting" + Name
+                        };
+                        res.render('admin/admin-error', details);
+                    }
+                });
+            })
+        }
+        else{
+            res.redirect('/admin/adminLogin');
+        }
     },
 
     searchProducts: function(req,res){
@@ -275,28 +305,34 @@ var productController = {
 
             if(req.session.userId) loggedIn = true;
             else loggedIn = false;
-            if(result != null) {
-                const details = {
-                    result,
-                    title: "Admin | Admin Products",
-                    loggedIn: loggedIn,
-                    userId: req.session.userId,
-                    username: req.session.adminUsername,
-                    error: null,
-                    path
-                };
-                res.render('admin/admin-product', details);
+
+            if(loggedIn){
+                if(result != null) {
+                    const details = {
+                        result,
+                        title: "Admin | Admin Products",
+                        loggedIn: loggedIn,
+                        userId: req.session.userId,
+                        username: req.session.adminUsername,
+                        error: null,
+                        path
+                    };
+                    res.render('admin/admin-product', details);
+                }
+                else {
+                    const details = {
+                        result,
+                        title: "Baked Goods | Error",
+                        loggedIn: loggedIn,
+                        userId: req.session.userId,
+                        username: req.session.adminUsername,
+                        error: "Oops! something went wrong with searching for the products"
+                    };
+                    res.render('admin/admin-error', details);
+                }
             }
-            else {
-                const details = {
-                    result,
-                    title: "Baked Goods | Error",
-                    loggedIn: loggedIn,
-                    userId: req.session.userId,
-                    username: req.session.adminUsername,
-                    error: "Oops! something went wrong with searching for the products"
-                };
-                res.render('admin/admin-error', details);
+            else{
+                res.redirect('/admin/adminLogin');
             }
         });
     }
